First attempt at triaging build errors.
This is a bit of a jumble and will need to be cleaned up and
split out later. General overview:

- Adds folly sub-modules which have been added since the original
  pull request for CMake support, incl. futures.
- Starts creating a very basic platform abstraction. Currently
  just has some basic definitions, support for Windows thread-local
  (actually, fiber-local) storage, and the beginnings of compat
  code for File.

There's a couple of features which I've simply butchered due to
bugs in VC++ that don't seem to have any simple workaround. In
particular I'm not 100% sure how to get the generic pointer argument
formatter in Format-inl.h to work due to the use of sizeof inside
std::enable_if. MS Connect says it's "fixed in the next version",
but it's not fixed in 2015 RTM.